kb fix

+++ include/keyboard.h
@@ -4,7 +4,7 @@
 #define KB_DATA_PORT  0x60
 #include <stdint.h>
 
-uint8_t kb_getkey();
+int kb_getkey();
 char kb_getchar();
 char kb_sctochar(int scancode);
